Fit description into available space in Look Up
Definitions in Look Up have only a bit of space available to them; if
they are larger, a "more" link is shown after only a partial
definition. If en element (such as <h1> in our case) doesn't fit into
this limited space, nothing is shown at  all.

Make the emoji and the description small enough to fit comfortably, as
well as look consistent with other definitions.

Fixes #5.